UBS-PI/PrathamWebApp_v2/requirements.txt
azure-cognitiveservices-speech==1.17.0
azure-common==1.1.27
azure-mgmt-cognitiveservices==3.0.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.3
boto==2.49.0
boto3==1.11.13
botocore==1.14.13
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cryptography==2.1.4
cvlib==0.2.6
cycler==0.10.0
Cython==0.29.21
Django==3.1.5
django-crispy-forms==1.10.0
django-extensions==3.1.3
dlib==19.22.0
docutils==0.15.2
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
futures==3.1.1
gast==0.3.3
gensim==3.8.1
getch==1.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
goslate==1.5.1
grpcio==1.29.0
h5py==2.10.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==1.6.0
imutils==0.5.4
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==1.0.0
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lxml==4.6.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
moviepy==1.0.3
msrest==0.6.21
msrestazure==0.6.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
nltk==3.5
notebook==6.0.3
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.5.1.48
packaging==20.8
pandas==0.25.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pio==0.0.3
pooch==1.3.0
proglog==0.1.9
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyDictionary==2.0.1
pygame==1.9.6
Pygments==2.6.1
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.0
pyrsistent==0.15.7
python-dateutil==2.8.1
python-debian==0.1.32
python-docx==0.8.10
pytz==2020.5
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
regex==2020.7.14
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.5.4
seaborn==0.9.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.13.2
six==1.15.0
smart-open==1.9.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
soupsieve==2.0.1
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
wadllib==1.3.2
wavio==0.0.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
zipp==3.1.0

